+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `summary.mcmc_fit` now displays the number of remaining parameters correctly.
 
+* Added the `as_mcmc.list()` function to convert MCMC fits to {coda} `mcmc.list` objects.
+
 # rater 1.0.0
 
 * Initial release